requirements.txt
pyyaml==5.2
taskcluster==24.1.0
